backend/requirements.txt
fastapi[all]
uvicorn[standard]
pydantic>=2.0
google-auth-oauthlib
google-api-python-client
google-generativeai
pymongo
python-dotenv
cryptography
